errors: move translation logic into module
Just moving code around so that triagebot can ping relevant parties when
translation logic is modified.
